Split classes from trace_link.py into separate files for improved modularity
